NewsAggregatorService/requirements.txt
fastapi
requests
uvicorn
google-generativeai
